requirements.txt
google-genai>=1.0.0
httpx>=0.27.0
Pillow>=10.0.0
python-dotenv>=1.0.0
gTTS>=2.5.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.7
